Increase HTTP and USB packet size.
Previously, we used a 4KiB packet size for USB transfers and HTTP
packets. This should work fine, but it's slow, especially for large IPP
print and eSCL scan jobs. Therefore, change it to 32KiB to speed things
up a bit.

As a data point, a scan on a virtual test printer was taking 1.8 seconds
with ipp-usb (which uses a 32KiB buffer size), and 8.5 seconds with
ippusbxd's 4 KiB buffer.

This could potentially be increased further; 32 KiB is not necessarily
the point of diminishing returns.