usb: fix race condition in usb_conn_acquire
If we do not have an available usb connection at the start of
usb_conn_acquire, we sleep and poll for several seconds until a
connection becomes available.

However, this check is done without holding the pool_manage_lock. This
allows for a race condition where we read num_avail and get a non-zero
value, but then some other thread acquires the pool_manage_lock first
and claims that free interface, causing us to error out when we then try
to claim the interface.

To fix this, acquire the pool_manage_lock before checking num_avail. This way,
when we exit our poll loop that waits for an available interface, we will be
certain that interface will continue to be available until we claim it.